Crate narwhal_test_utils

Modules

cluster

Macros

test_channel

test_committed_certificates_channel

test_new_certificates_channel

Structs

AuthorityFixture

Builder

CommitteeFixture

PrimaryToPrimaryMockServer

PrimaryToWorkerMockServer

TestLeaderConfiguration

WorkerFixture

WorkerToWorkerMockServer

Enums

TestLeaderSupport

Constants

CERTIFICATES_CF

CERTIFICATE_DIGEST_BY_ORIGIN_CF

CERTIFICATE_DIGEST_BY_ROUND_CF

HEADERS_CF

PAYLOAD_CF

VOTES_CF

Functions

batch

Batches

batch_with_rand

batch_with_transactions

batches

generate multiple fixture batches. The number of generated batches are dictated by the parameter num_of_batches.

create_batch_store

ensure_test_environment

fixture_batch_with_transactions

fixture_payload

Headers, Votes, Certificates

fixture_payload_with_rand

get_protocol_config

latest_protocol_version

make_certificates

make_certificates_with_epoch

make_certificates_with_leader_configuration

Creates fully connected DAG for the dictated rounds but with specific conditions for the leaders. By providing the leader_configuration we can dictate the setup for specific leaders of specific rounds. For a leader the following can be configured:

make_certificates_with_slow_nodes

make_optimal_certificates

make_optimal_signed_certificates

make_signed_certificates

mock_certificate

mock_certificate_with_epoch

mock_certificate_with_rand

mock_signed_certificate

random_key

Keys, Committee

random_network

temp_dir

test_network

this_cert_parents_with_slow_nodes

transaction

transaction_with_rand